+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Adapt to new pwix-ui-layout
+    - Replace bootstrap and @popperjs/core npm dependencies with pwix:ui-boostrap5
 
 ### 2.4.0